reorganize the json file search order

+++ openimu/openimu.py
@@ -99,28 +99,38 @@ class OpenIMU:
             print('creat data folder for store measure data in future')
             os.makedirs("data")
             
-        if not os.path.exists('app_config'):
-            logging.debug("start to download the JSON files")  
-            print('downloading config json files from github, please waiting for a while')
-            os.makedirs('app_config')
-            for app_name in get_app_names():
-                os.makedirs('app_config'+ '/' + app_name)
-            i = 0
-            for url in get_app_urls():
-                filepath = 'app_config' + '/' + get_app_names()[i] + '/' + 'openimu.json'
-                i= i+1
-                try:
-                    r = requests.get(url) 
-                    # r = requests.session().get(url)
-                    with open(filepath, "wb") as code:
-                        code.write(r.content)     
-                except Exception as e:
-                    logging.info("downloading the JSON file failed, url: {0}".format(url)) 
+        # check the local json file openimu.json, if no, download the app json files from github.
+        if not os.path.isfile("./python-openimu/openimu.json"): 
+            local_jsonfile_flag = False # no local json file
+            if not os.path.exists('app_config'):
+                print('downloading config json files from github, please waiting for a while')
+                os.makedirs('app_config')
+                for app_name in get_app_names():
+                    os.makedirs('app_config'+ '/' + app_name)
+                i = 0
+                for url in get_app_urls():
+                    filepath = 'app_config' + '/' + get_app_names()[i] + '/' + 'openimu.json'
+                    i= i+1
+                    try:
+                        r = requests.get(url) 
+                        # r = requests.session().get(url)
+                        with open(filepath, "wb") as code:
+                            code.write(r.content)     
+                    except Exception as e:
+                        logging.info("downloading the JSON file failed, url: {0}".format(url)) 
+            else:
+                print('the config json files have already been downloaded from github')
 
-                    
-        # Load the basic openimu.json(IMU application)
-        with open('app_config/IMU/openimu.json') as json_data:
-            self.imu_properties = json.load(json_data)
+            try:
+                # Load the basic app_config openimu.json(IMU application)
+                with open('app_config/IMU/openimu.json') as json_data:
+                    self.imu_properties = json.load(json_data)
+            except Exception as e:
+                print(e)
+        else:
+            local_jsonfile_flag = True #has json file
+            with open('./python-openimu/openimu.json') as json_data:
+                self.imu_properties = json.load(json_data)
 
     def find_device(self):
         ''' Finds active ports and then autobauds units
